crawler/requirements.txt
# requirements.txt - Simplified version

# Core dependencies
requests==2.31.0
requests[socks]==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
psycopg2-binary==2.9.9
redis==5.0.1

# Distributed processing
celery==5.3.4
backoff==2.2.1

# Elasticsearch
elasticsearch==8.11.0

# GeoIP & ASN Enrichment
geoip2==4.8.0
maxminddb==2.8.2

# Utils
python-dotenv==1.0.0
